requirements.txt
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
class-registry==2.1.2
filters==1.3.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==2.0.6
